Add hashed and private read-write set to proto
This CR adds protobuf struct for hashed read/write set and private
read/write set.

To the existing KVRWSet, i have added a field to store hashed read write
set of all collections. For a private read, we can store the hash in key
field but keep the version as is. For private writes, we can store hash
of key and hash of value but store the IsDelete as is. Hence, reused
existing KVRead and KVWrite.

For the range query info, it does not make sense to store the hash of
each field in the RangeQueryInfo struct. Note that we cannot do a range
query on state DB using hashed startKey and endKey to find phantom
reads.

Instead of hashed RangeQueryInfo per collection, i have added a string
field to hold a hash of entire private RangeQueryInfos of that collection.
During validation, we can compare the hash of entire private RangeQueryInfo
present in the tempDB with the hash present in the block. I thought of
doing the same for hashed write-set too as we may not store this value in stateDB.
However, it might be useful during validation phase (while comparing the
write set of Ti with read set of Ti+1 -- especially when the private
write-set is unavailable/not received).

If the peer is configured to move on without private write-set
(implicit: phantom read check is disabled), we must not use
RangeQueryInfo. Instead add the read <key, version> to private read
set and hashed read set.

I have also added PrivateCollectionsKVRWSet to hold private read-write
set of all collections. For clear differentiation on variable names, i
have used Reads, HashedReads, and PrivateReads (similarly for the Writes).